requirements.txt
alembic==1.7.3
certifi==2021.5.30
charset-normalizer==2.0.6
clearbit==0.1.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyhunter==1.7
PyJWT==2.1.0
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
